Lightcurve fix on lightcurve.py for missing key
I had taken lightcurve.py at the latest version but this is now a change on the orginal lightcurve.py in tutorial. It generates fine. The only peculiarity is that a lightcurve object is not  returned from lightcurve method. However the lightcurve is saved fine and a reload with appropriate casting allows Max's plot code to be used - will amend code snippet in tutorial 3 to match up